Disable Valheim
The whole point of `@zeroindexed/valheim` and `@zeroindexed/valheimctl`
was to provide the cheapest possible way to run a Valheim server, with a
web UI to start the server on demand and automated shutdown when idle.

All the pieces worked great except for the `LoadBalancer` resource,
which is always online even though the service it routes traffic to has
no active pods.

This ends up costing ~$18/month, which breaks the "cheapest possible"
promise. Doing away with the `LoadBalancer` is going to require
significant rethinking of how this is deployed.